LayoutTests/platform/chromium/fast/dom/xmlhttprequest-constructor-in-detached-document-expected.txt

CONSOLE MESSAGE: line 12: Uncaught ReferenceError: XMLHttpRequest constructor's associated context is not available
Text for bug 25290: Crash when constructing XMLHttpRequest in a detached document.

PASS
